+++ rpcs3/Emu/SysCalls/Modules/cellGcmSys.h
@@ -30,7 +30,7 @@ void cellGcmGetOffsetTable(mem_ptr_t<CellGcmOffsetTable> table);
 s32 cellGcmIoOffsetToAddress(u32 ioOffset, u64 address);
 s32 cellGcmMapEaIoAddress(u32 ea, u32 io, u32 size);
 s32 cellGcmMapEaIoAddressWithFlags(u32 ea, u32 io, u32 size, u32 flags);
-s32 cellGcmMapMainMemory(u64 ea, u32 size, mem32_t offset);
+s32 cellGcmMapMainMemory(u32 ea, u32 size, mem32_t offset);
 s32 cellGcmReserveIoMapSize(u32 size);
 s32 cellGcmUnmapEaIoAddress(u64 ea);
 s32 cellGcmUnmapIoAddress(u64 io);
